requirements.txt
# Core dependencies
textual>=0.60.0
httpx>=0.27.0
httpx-sse>="0.4.3"

# Configuration / data modeling
pydantic>=2.0
python-dotenv>=1.0

# Developer tools
black>=24.3.0
ruff>=0.5.0
mypy>=1.8.0
pytest>=8.2.0
pytest-asyncio>=0.23.5
